+++ uefi-test-runner/src/runtime/vars.rs
@@ -1,6 +1,6 @@
 use log::info;
 use uefi::prelude::*;
-use uefi::table::runtime::{VariableAttributes, VariableVendor};
+use uefi::runtime::{VariableAttributes, VariableVendor};
 use uefi::{guid, runtime, CStr16, Error};
 
 /// Test variable name.
@@ -16,63 +16,8 @@ const VALUE: &[u8] = b"TestValue";
 const ATTRS: VariableAttributes =
     VariableAttributes::BOOTSERVICE_ACCESS.union(VariableAttributes::RUNTIME_ACCESS);
 
-fn test_variables(rt: &RuntimeServices) {
-    info!("Testing set_variable");
-    rt.set_variable(NAME, VENDOR, ATTRS, VALUE)
-        .expect("failed to set variable");
-
-    info!("Testing get_variable_size");
-    let size = rt
-        .get_variable_size(NAME, VENDOR)
-        .expect("failed to get variable size");
-    assert_eq!(size, VALUE.len());
-
-    info!("Testing get_variable");
-    let mut buf = [0u8; 9];
-    let (data, attrs) = rt
-        .get_variable(NAME, VENDOR, &mut buf)
-        .expect("failed to get variable");
-    assert_eq!(data, VALUE);
-    assert_eq!(attrs, ATTRS);
-
-    info!("Testing get_variable_boxed");
-    let (data, attrs) = rt
-        .get_variable_boxed(NAME, VENDOR)
-        .expect("failed to get variable");
-    assert_eq!(&*data, VALUE);
-    assert_eq!(attrs, ATTRS);
-
-    info!("Testing variable_keys");
-    let variable_keys = rt.variable_keys().expect("failed to get variable keys");
-    info!("Found {} variables", variable_keys.len());
-    // There are likely a bunch of variables, only print out the first one
-    // during the test to avoid spamming the log.
-    if let Some(key) = variable_keys.first() {
-        info!("First variable: {}", key);
-    }
-
-    // Test that the `runtime::variable_keys` iterator gives exactly the same
-    // list as the `RuntimeServices::variable_keys` function.
-    assert_eq!(
-        runtime::variable_keys()
-            .map(|k| k.unwrap())
-            .collect::<alloc::vec::Vec<_>>(),
-        variable_keys
-    );
-
-    info!("Testing delete_variable()");
-    rt.delete_variable(NAME, VENDOR)
-        .expect("failed to delete variable");
-    assert_eq!(
-        rt.get_variable(NAME, VENDOR, &mut buf)
-            .unwrap_err()
-            .status(),
-        Status::NOT_FOUND
-    );
-}
-
 /// Test the variable functions in `uefi::runtime`.
-fn test_variables_freestanding() {
+fn test_variables() {
     assert!(!runtime::variable_exists(NAME, VENDOR).unwrap());
 
     // Create the test variable.
@@ -121,20 +66,17 @@ fn test_variables_freestanding() {
     assert!(!find_by_key());
 }
 
-fn test_variable_info(rt: &RuntimeServices) {
+fn test_variable_info() {
     let attr = VariableAttributes::BOOTSERVICE_ACCESS | VariableAttributes::NON_VOLATILE;
-    let info = rt.query_variable_info(attr).unwrap();
+    let info = runtime::query_variable_info(attr).unwrap();
     info!("Storage for non-volatile boot-services variables: {info:?}");
-    assert_eq!(info, runtime::query_variable_info(attr).unwrap());
 
     let attr = VariableAttributes::BOOTSERVICE_ACCESS | VariableAttributes::RUNTIME_ACCESS;
-    let info = rt.query_variable_info(attr).unwrap();
+    let info = runtime::query_variable_info(attr).unwrap();
     info!("Storage for volatile runtime variables: {info:?}");
-    assert_eq!(info, runtime::query_variable_info(attr).unwrap());
 }
 
-pub fn test(rt: &RuntimeServices) {
-    test_variables(rt);
-    test_variable_info(rt);
-    test_variables_freestanding();
+pub fn test() {
+    test_variable_info();
+    test_variables();
 }