+++ uefi/src/table/mod.rs
@@ -2,14 +2,13 @@
 
 pub mod boot;
 pub mod cfg;
-pub mod runtime;
 
 mod header;
 mod system;
 
 pub use header::Header;
 #[allow(deprecated)]
-pub use system::{Boot, Runtime, SystemTable};
+pub use system::{Boot, SystemTable};
 pub use uefi_raw::table::Revision;
 
 use core::ptr::{self, NonNull};
@@ -75,25 +74,6 @@ pub fn system_table_boot() -> Option<SystemTable<Boot>> {
             None
         } else {
             Some(SystemTable::<Boot>::from_ptr(st.cast()).unwrap())
-        }
-    }
-}
-
-/// Get the system table while runtime services are active.
-#[deprecated = "Use the uefi::runtime module instead. See https://github.com/rust-osdev/uefi-rs/blob/HEAD/docs/funcs_migration.md"]
-#[allow(deprecated)]
-pub fn system_table_runtime() -> Option<SystemTable<Runtime>> {
-    let st = SYSTEM_TABLE.load(Ordering::Acquire);
-    if st.is_null() {
-        return None;
-    }
-
-    // SAFETY: the system table is valid per the requirements of `set_system_table`.
-    unsafe {
-        if (*st).runtime_services.is_null() {
-            None
-        } else {
-            Some(SystemTable::<Runtime>::from_ptr(st.cast()).unwrap())
         }
     }
 }